+++ packages/jaspr/test/framework/inherited_model/inherited_model_test.dart
@@ -20,7 +20,7 @@ void testAgainstDto(MyDto src, {required Set<Key> expectedRebuilds}) {
 void main() {
   group('inherited model test', () {
     testComponents('should update dependencies only', (tester) async {
-      var controller = await tester.pumpTestComponent(App());
+      var controller = tester.pumpTestComponent(App());
       Future<void> rebuildThenTestAgainstDto(
         MyDto src, {
         required Set<Key> expectedRebuilds,