+++ packages/jaspr_riverpod/test/context_watch_test.dart
@@ -10,7 +10,7 @@ final autoDisposeCounter = StateProvider.autoDispose((ref) => 0);
 void main() {
   group('context.watch', () {
     testComponents('returns provider state and rebuilds on change', (tester) async {
-      await tester.pumpComponent(providerApp((context) sync* {
+      tester.pumpComponent(providerApp((context) sync* {
         yield Button(
           label: '${context.watch(counter)}',
           onPressed: () {
@@ -28,7 +28,7 @@ void main() {
     });
 
     testComponents('returns overridden provider state', (tester) async {
-      await tester.pumpComponent(providerApp((context) sync* {
+      tester.pumpComponent(providerApp((context) sync* {
         yield Builder(builder: (context) sync* {
           yield Button(
             key: const ValueKey('a'),
@@ -69,7 +69,7 @@ void main() {
     });
 
     testComponents('provider is autodisposed when no longer watched', (tester) async {
-      await tester.pumpComponent(providerApp((context) sync* {
+      tester.pumpComponent(providerApp((context) sync* {
         var showCounter = true;
         yield StatefulBuilder(
           builder: (context, setState) sync* {